+++ src/ResourceManager/Websites/Commands.Websites/ChangeLog.md
@@ -19,6 +19,12 @@
 -->
 ## Current Release
 
+## Version 5.1.0
+* Updating to use the latest .NET SDK version (2.0.0)
+* Added two new cmdlets: Get-AzureRmDeletedWebApp and Restore-AzureRmDeletedWebApp
+* New-AzureRmAppServicePlan -HyperV switch is added for create app service plan with windows container
+* New-AzureRmWebApp/ New-AzureRmWebAppSlot/ Set-AzureRmWebApp/ Set-AzureRmWebAppSlot - New parameters (–ContainerRegistryUser string -ContainerRegistryPassword secureString -EnableContainerContinuousDeployment) added for creating and managing windows container app
+
 ## Version 5.0.9
 * Fixed issue with default resource groups not being set.
 * Updated common runtime assemblies